pyproject.toml
[project]
name = "clauderfall"
version = "0.1.0"
description = "Precision context system for code-generating agents."
readme = "docs/README.md"
requires-python = ">=3.12"
dependencies = [
  "pydantic>=2.11,<3",
  "PyYAML>=6.0,<7",
  "sqlalchemy>=2.0,<3",
  "typer>=0.16,<1",
]

[build-system]
requires = ["hatchling>=1.27"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[dependency-groups]
dev = [
  "pytest>=8.3,<9",
  "pytest-cov>=6.1,<7",
]

[tool.hatch.build.targets.wheel]
packages = ["src/clauderfall"]

[tool.pytest.ini_options]
testpaths = ["tests"]
